requirements.txt
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
Django==2.1
django-allauth==0.37.1
django-ckeditor==5.6.1
django-js-asset==1.1.0
idna==2.7
oauthlib==2.1.0
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
urllib3==1.23
